Updates GeneConversionTest
