+++ build.gradle
@@ -52,6 +52,9 @@ dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter:5.8.2'
     // https://mvnrepository.com/artifact/io.projectreactor/reactor-test
     testImplementation 'io.projectreactor:reactor-test:3.4.16'
+    // Assertions
+    // https://mvnrepository.com/artifact/org.assertj/assertj-core
+    testImplementation 'org.assertj:assertj-core:3.22.0'
     // https://mvnrepository.com/artifact/org.awaitility/awaitility
     testImplementation 'org.awaitility:awaitility:4.2.0'
 }
